Add lab 4

+++ info_sec_lab4_RSA/utils.cpp
@@ -14,13 +14,13 @@ string _uintToBin(uint64_t n)
 	return result;
 }
 
-uint32_t _binStrToUlong(const string& str)
+uint64_t _binStrToUlong(const string& str)
 {
-	uint32_t res = 0;
-	for (int32_t i = str.size() - 1, j = 0; i >= 0; --i, ++j)
+	uint64_t res = 0;
+	for (int64_t i = str.size() - 1, j = 0; i >= 0; --i, ++j)
 	{
 //		uint32_t factor = strtol(str.data() + j, nullptr, 10);
-		uint32_t factor = (str[j] == '1') ? 1 : 0;
+		uint64_t factor = (str[j] == '1') ? 1 : 0;
 		res += factor * pow(2, i);
 	}
 	return res;
@@ -43,8 +43,8 @@ string supplement(string str, uint64_t size)
 string _binStrToSymbols(const string& bitstr, uint64_t symb_bits_size)
 {
 	string res;
-	uint16_t n = bitstr.size() / (sizeof (char) * 8);
-	for (uint16_t i = 0; i < n; ++i)
+	uint64_t n = bitstr.size() / (sizeof (char) * 8);
+	for (uint64_t i = 0; i < n; ++i)
 	{
 		string substr = bitstr.substr(i * symb_bits_size, symb_bits_size);
 		uint8_t ch = _binStrToUlong(substr);
